Add input files for leap
1stp-prepped was prepared using MOE. Capping groups were used, since
residues are missing at the ends. The structure was protonated to pH 7.4
(no charged histidines). Hydrogen atoms on ACE and BTN (biotin) were
removed in VIM, and the atom naming was changed to be consistent with
the prep file.
The prep and frcmod files were generated with the standard antechamber
procedure (Gaff2, RESP/HF charges). The input structure was downloaded
from the PDB ("Model sdf" from https://www.rcsb.org/ligand/BTN)